requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
yfinance>=0.2.31
numpy
pandas
scikit-learn>=1.4.0
python-multipart>=0.0.6
httpx>=0.25.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
reportlab>=4.0.4
pytest-cov>=4.1.0
weasyprint>=60.2
prometheus-client>=0.19.0
schedule>=1.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
textblob>=0.17.1
newsapi-python>=0.2.7
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
tenacity>=8.2.3
cachetools>=5.3.2 